Btrfs-progs: optimize function btrfs_read_chunk_tree
After reading all device items from the chunk tree, don't
exit the loop and then navigate down the tree again to find
the chunk items. Instead, if while reading the device items
we find a chunk item in the leaf keep iterating over the leaf
and process the chunk items - at the moment chunk items always
follow immediately the device items in the chunk tree, and if
this fact changes in the future, revert back to previous
behaviour of exiting the loop and navigate down the tree again
to search for chunk items, but use BTRFS_FIRST_CHUNK_TREE_OBJECTID
as the object id for the search key, as any chunk item has a
key that has this object id.

This is a port of the corresponding kernel patch to keep both
kernel and btrfs-progs identical:

https://patchwork.kernel.org/patch/2816401/

Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>